requirements.txt
flask
scikit-learn
pandas
numpy
PyYAML